Move route to config
