+++ src/components/EntriesTable/index.tsx
@@ -41,7 +41,7 @@ import {
     DeleteEntryMutationVariables,
 } from '#generated/types';
 
-import route from '../../Root/App/Multiplexer/route';
+import route from '#config/routes';
 import styles from './styles.css';
 
 interface Entity {
