+++ src/views/Crises/index.tsx
@@ -43,7 +43,7 @@ import {
     DeleteCrisisMutationVariables,
 } from '#generated/types';
 
-import route from '../../Root/App/Multiplexer/route';
+import route from '#config/routes';
 import styles from './styles.css';
 
 interface Entity {
